+++ bin/plugin-manager-cli.js
@@ -4,8 +4,8 @@
  * @tagline         Plugin management commands for jPulse Framework
  * @description     Handles plugin install, update, remove, enable, disable, list, info, publish
  * @file            bin/plugin-manager-cli.js
- * @version         1.4.11
- * @release         2026-01-11
+ * @version         2.4.12
+ * @release         2026-01-12
  * @repository      https://github.com/jpulse-net/jpulse-framework
  * @author          Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/
  * @copyright       2025 Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/