+++ plugins/hello-world/webapp/view/jpulse-navigation.js
@@ -3,8 +3,8 @@
  * @tagline         Navigation of the Hello World Plugin
  * @description     Navigation for the Hello World Plugin, appended to the framework navigation
  * @file            plugins/hello-world/webapp/view/jpulse-navigation.js
- * @version         1.4.11
- * @release         2026-01-11
+ * @version         2.4.12
+ * @release         2026-01-12
  * @repository      https://github.com/jpulse-net/jpulse-framework
  * @author          Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/
  * @copyright       2025 Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/